fix bug#1025014 - remove unuseds, quell signed/unsigned warnings in 5.2.pre1
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@11048 06827809-a52a-0410-b366-d66718629ded

+++ agent/helpers/old_api.c
@@ -125,7 +125,7 @@ netsnmp_register_mib_table_row(const char *moduleName,
         struct variable *vr =
             (struct variable *) ((char *) var + (i * varsize));
         netsnmp_handler_registration *r;
-        if ( var_subid > mibloclen ) {
+        if ( var_subid > (int)mibloclen ) {
             break;    /* doesn't make sense */
         }
         r = SNMP_MALLOC_TYPEDEF(netsnmp_handler_registration);
